requirements.txt
streamlit>=1.24.0
pandas>=1.5.3
numpy>=1.24.3
plotly>=5.15.0
pillow>=9.5.0
uuid>=1.30 